+++ backend/windmill-queue/src/schedule.rs
@@ -11,6 +11,7 @@ use crate::QueueTransaction;
 use sqlx::{query_scalar, Postgres, Transaction};
 use std::str::FromStr;
 use windmill_common::jobs::JobPayload;
+use windmill_common::schedule::schedule_to_user;
 use windmill_common::{
     error::{self, Result},
     schedule::Schedule,
@@ -146,8 +147,4 @@ pub async fn exists_schedule(
     .unwrap_or(false);
 
     Ok(exists)
-}
-
-fn schedule_to_user(path: &str) -> String {
-    format!("schedule-{}", path.replace('/', "-"))
 }