+++ frontend/src/lib/components/flows/previousResults.ts
@@ -109,7 +109,12 @@ export function getStepPropPicker(
 	args: any,
 	include_node: boolean
 ): StepPropPicker {
-	const flowInput = getFlowInput(dfs(parentModule?.id, flow), flowState, args, flow.schema)
+	const flowInput = getFlowInput(
+		dfs(parentModule?.id, flow),
+		flowState,
+		args,
+		flow.schema as Schema
+	)
 
 	const previousIds = dfs(id, flow, false)
 		.map((x) => {